requirements.txt
fastapi
uvicorn[standard]
google-generativeai
resend
python-dotenv
pydantic[email]
